Add ManagerChannel permission for update/remove welcome channel #2

+++ src/main/kotlin/core/extensions/command/slash/moderation/RemoveWelcomeChannelCommandExtension.kt
@@ -5,7 +5,9 @@ import com.kotlindiscord.kord.extensions.commands.application.slash.publicSubCom
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.publicSlashCommand
 import com.kotlindiscord.kord.extensions.types.respond
+import com.kotlindiscord.kord.extensions.utils.hasPermission
 import core.common.Common
+import dev.kord.common.entity.Permission
 import dev.kord.core.kordLogger
 import domain.repository.guild.GuildRepository
 import org.koin.core.component.inject
@@ -31,16 +33,25 @@ class RemoveWelcomeChannelCommandExtension : Extension() {
                     name = "welcome"
                     description = "Remove welcome channel from database"
 
+                    requirePermission(Permission.ManageChannels)
+                    allowInDms = false
+
                     action {
                         try {
+                            val member = member?.asMember() ?: throw NullPointerException("Member is null")
+
+                            if (!member.hasPermission(Permission.ManageChannels))
+                                throw RuntimeException("User does not have manage channels permission for this command!")
+
                             val guildId = guild?.id?.value?.toLong() ?: throw NullPointerException("Guild is null")
+
                             guildRepository.removeWelcomeMessageChannel(guildId)
 
                             respond {
                                 content = "Welcome channel removed!"
                             }
                         } catch (e: Exception) {
-                            kordLogger.error(e.localizedMessage)
+                            kordLogger.error("RemoveWelcomeChannel: ${e.localizedMessage}")
 
                             respond {
                                 content = "**Error:** ${e.localizedMessage}"
